FINAL/chess_ai/requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Chess engine
python-chess>=1.999

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Progress / system utils
tqdm>=4.65.0
psutil>=5.9.0

# Experiment / evaluation tooling
kaggle>=1.5.13
bayeselo>=0.5.0  # if you use a Python wrapper, else remove

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image processing (for any board/piece assets)
Pillow>=9.5.0

# Backend + API
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
python-engineio>=4.5.1

# HTTP / download helpers (for model from Drive, etc.)
requests>=2.31.0
gdown>=4.7.1
